backend/requirements.txt

# ========== CORE FRAMEWORK ==========
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
# ========== PDF/DOCX PROCESSING ==========
PyPDF2==3.0.1           
python-docx==1.1.0
# ========== AI/ML ==========
google-generativeai==0.3.2

# ========== IMAGE PROCESSING ==========
Pillow==10.2.0

# ========== DATA PROCESSING ==========
pandas==2.1.4
numpy==1.26.3

# ========== HTTP/NETWORKING ==========
httpx==0.26.0
requests==2.31.0
aiofiles==23.2.1

# ========== MONITORING - OpenTelemetry (Core only) ==========
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
opentelemetry-instrumentation-requests


# ========== MONITORING - Prometheus (Direct) ==========
prometheus-client==0.19.0

# ========== LOGGING ==========
python-json-logger==2.0.7

# ========== UTILITIES ==========
validators==0.22.0

# ========== TESTING (Optional) ==========
# pytest==7.4.4
# pytest-asyncio==0.23.3